games/AetherMon/requirements.txt
rich>=13.7.0
Pillow>=10.3.0
colorama>=0.4.6
requests>=2.31.0
pytest>=8.1.0
